clean up identity and contacts
