+++ src/service/contact_service.rs
@@ -5,7 +5,7 @@ use std::{collections::HashMap, sync::Arc};
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 
-use crate::{bill::identity::Identity, dht::Client, persistence::ContactStoreApi};
+use crate::{dht::Client, persistence::ContactStoreApi, service::identity_service::Identity};
 
 use super::Result;
