+++ src/web/handlers/identity.rs
@@ -1,7 +1,7 @@
 use super::super::data::IdentityForm;
 use crate::service::{self, Result};
 use crate::web::data::{ChangeIdentityForm, NodeId};
-use crate::{bill::identity::Identity, service::ServiceContext};
+use crate::{service::identity_service::Identity, service::ServiceContext};
 use libp2p::PeerId;
 use rocket::form::Form;
 use rocket::http::Status;
